ADM-440 fix: rename vo to dto in controller directory
