+++ backend/src/main/java/heartbeat/controller/board/dto/request/BoardType.java
@@ -1,4 +1,4 @@
-package heartbeat.controller.board.vo.request;
+package heartbeat.controller.board.dto.request;
 
 public enum BoardType {
