+++ backend/src/main/java/heartbeat/controller/board/dto/response/JiraColumnResponse.java
@@ -1,4 +1,4 @@
-package heartbeat.controller.board.vo.response;
+package heartbeat.controller.board.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
