+++ backend/src/main/java/heartbeat/controller/pipeline/dto/request/PipelineStepsParam.java
@@ -1,21 +1,23 @@
-package heartbeat.controller.pipeline.vo.request;
+package heartbeat.controller.pipeline.dto.request;
 
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PipelineParam {
+public class PipelineStepsParam {
 
-	@Valid
-	@NotBlank(message = "Token cannot be empty.")
-	private String token;
+	@NotBlank(message = "PipelineName is required")
+	private String pipelineName;
+
+	@NotBlank(message = "Repository is required")
+	private String repository;
+
+	@NotBlank(message = "OrgName is required")
+	private String orgName;
 
 	@NotBlank(message = "StartTime is required")
 	private String startTime;
