+++ backend/src/main/java/heartbeat/controller/pipeline/dto/response/PipelineStepsResponse.java
@@ -1,4 +1,4 @@
-package heartbeat.controller.pipeline.vo.response;
+package heartbeat.controller.pipeline.dto.response;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
