+++ backend/src/main/java/heartbeat/service/board/jira/JiraColumnResult.java
@@ -1,6 +1,6 @@
 package heartbeat.service.board.jira;
 
-import heartbeat.controller.board.vo.response.JiraColumnResponse;
+import heartbeat.controller.board.dto.response.JiraColumnResponse;
 import java.util.List;
 import lombok.Builder;
 import lombok.Data;
