+++ backend/src/main/java/heartbeat/service/board/jira/JiraService.java
@@ -11,12 +11,12 @@ import heartbeat.client.dto.board.jira.Issuetype;
 import heartbeat.client.dto.board.jira.JiraBoardConfigDTO;
 import heartbeat.client.dto.board.jira.JiraColumn;
 import heartbeat.client.dto.board.jira.StatusSelfDTO;
-import heartbeat.controller.board.vo.request.BoardRequestParam;
-import heartbeat.controller.board.vo.request.BoardType;
-import heartbeat.controller.board.vo.response.BoardConfigResponse;
-import heartbeat.controller.board.vo.response.ColumnValue;
-import heartbeat.controller.board.vo.response.JiraColumnResponse;
-import heartbeat.controller.board.vo.response.TargetField;
+import heartbeat.controller.board.dto.request.BoardRequestParam;
+import heartbeat.controller.board.dto.request.BoardType;
+import heartbeat.controller.board.dto.response.BoardConfigResponse;
+import heartbeat.controller.board.dto.response.ColumnValue;
+import heartbeat.controller.board.dto.response.JiraColumnResponse;
+import heartbeat.controller.board.dto.response.TargetField;
 import heartbeat.exception.RequestFailedException;
 import jakarta.annotation.PreDestroy;
 import lombok.RequiredArgsConstructor;
