+++ backend/src/main/java/heartbeat/service/pipeline/buildkite/BuildKiteService.java
@@ -6,12 +6,12 @@ import heartbeat.client.dto.pipeline.buildkite.BuildKiteBuildInfo;
 import heartbeat.client.dto.pipeline.buildkite.BuildKiteJob;
 import heartbeat.client.dto.pipeline.buildkite.BuildKiteOrganizationsInfo;
 import heartbeat.client.dto.pipeline.buildkite.BuildKiteTokenInfo;
-import heartbeat.controller.pipeline.vo.request.PipelineParam;
-import heartbeat.controller.pipeline.vo.request.PipelineStepsParam;
-import heartbeat.controller.pipeline.vo.response.BuildKiteResponse;
-import heartbeat.controller.pipeline.vo.response.Pipeline;
-import heartbeat.controller.pipeline.vo.response.PipelineStepsResponse;
-import heartbeat.controller.pipeline.vo.response.PipelineTransformer;
+import heartbeat.controller.pipeline.dto.request.PipelineParam;
+import heartbeat.controller.pipeline.dto.request.PipelineStepsParam;
+import heartbeat.controller.pipeline.dto.response.BuildKiteResponse;
+import heartbeat.controller.pipeline.dto.response.Pipeline;
+import heartbeat.controller.pipeline.dto.response.PipelineStepsResponse;
+import heartbeat.controller.pipeline.dto.response.PipelineTransformer;
 import heartbeat.exception.PermissionDenyException;
 import heartbeat.exception.RequestFailedException;
 import heartbeat.util.TokenUtil;
