+++ backend/src/main/java/heartbeat/util/StringToEnumConverter.java
@@ -1,6 +1,6 @@
 package heartbeat.util;
 
-import heartbeat.controller.board.vo.request.BoardType;
+import heartbeat.controller.board.dto.request.BoardType;
 import org.springframework.core.convert.converter.Converter;
 
 public class StringToEnumConverter implements Converter<String, BoardType> {
