+++ backend/src/test/java/heartbeat/service/pipeline/buildkite/BuildKiteServiceTest.java
@@ -9,11 +9,11 @@ import heartbeat.client.dto.pipeline.buildkite.BuildKiteJob;
 import heartbeat.client.dto.pipeline.buildkite.BuildKiteOrganizationsInfo;
 import heartbeat.client.dto.pipeline.buildkite.BuildKitePipelineDTO;
 import heartbeat.client.dto.pipeline.buildkite.BuildKiteTokenInfo;
-import heartbeat.controller.pipeline.vo.request.PipelineParam;
-import heartbeat.controller.pipeline.vo.request.PipelineStepsParam;
-import heartbeat.controller.pipeline.vo.response.BuildKiteResponse;
-import heartbeat.controller.pipeline.vo.response.Pipeline;
-import heartbeat.controller.pipeline.vo.response.PipelineStepsResponse;
+import heartbeat.controller.pipeline.dto.request.PipelineParam;
+import heartbeat.controller.pipeline.dto.request.PipelineStepsParam;
+import heartbeat.controller.pipeline.dto.response.BuildKiteResponse;
+import heartbeat.controller.pipeline.dto.response.Pipeline;
+import heartbeat.controller.pipeline.dto.response.PipelineStepsResponse;
 import heartbeat.exception.RequestFailedException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
